Essence::replace should not check/replace already linked URLs
When using Essence::replace on a piece of text, it should embed
plaintext links. But if the URL is inside a text link, it is not
a good idea to embed it, both in regards to invalidation of the
HTML, and to the intent of the user who wrote that piece of text.

Added a new test method: EssenceTest::testDontReplaceLinks

The change was basically merely changing \b to \s in the Daring
Fireball URL regex so that it would only match URLs surrounded
by whitespace (as compared to surrounded by non-word characters
like quotes).